+++ cron/stack_part_stock_in_paid_orders.php
@@ -12,7 +12,7 @@
 
 require_once 'common.php';
 
-$print_est = true;
+$print_est = false;
 
 
 $editor = array(
@@ -45,11 +45,13 @@ if ($result = $db->query($sql)) {
 
 
 $lap_time0 = date('U');
+$lap_time1= date('U');
+
 $contador  = 0;
 
 
 $sql = sprintf(
-    "SELECT `Stack Key`,`Stack Object Key` FROM `Stack Dimension`  where `Stack Operation`='part_stock_in_paid_orders' "
+    "SELECT `Stack Key`,`Stack Object Key` FROM `Stack Dimension`  where `Stack Operation`='part_stock_in_paid_orders' ORDER BY RAND()"
 );
 
 if ($result = $db->query($sql)) {
@@ -58,15 +60,26 @@ if ($result = $db->query($sql)) {
 
         if($part->id){
 
-            $editor['Date'] = gmdate('Y-m-d H:i:s');
-            $part->editor = $editor;
 
-            $part->update_stock_in_paid_orders();
+            $sql=sprintf('select `Stack Key` from `Stack Dimension` where `Stack Key`=%d ',$row['Stack Key']);
+
+            if ($result2=$db->query($sql)) {
+                if ($row2 = $result2->fetch()) {
+                    $sql=sprintf('delete from `Stack Dimension`  where `Stack Key`=%d ',$row['Stack Key']);
+                    $db->exec($sql);
+
+                    $editor['Date'] = gmdate('Y-m-d H:i:s');
+                    $part->editor = $editor;
+
+                    $part->update_stock_in_paid_orders();
 
 
 
-            $sql=sprintf('delete from `Stack Dimension`  where `Stack Key`=%d ',$row['Stack Key']);
-            $db->exec($sql);
+
+                }
+            }
+
+
         }
 
         $contador++;
@@ -84,6 +97,8 @@ if ($result = $db->query($sql)) {
     print_r($error_info = $db->errorInfo());
     exit;
 }
-
+if($total>0){
+    printf("%s: xP  %s/%s %.2f min \n",gmdate('Y-m-d H:i:s'),$contador,$total,($lap_time1 - $lap_time0)/60);
+}
 
 ?>
